Added support for debitor's structured address data (street name, post code, etc.)
Currently the direct debit transfer DOM builder only supports the free format address line (`AdrLine`) which is not accepted by our bank, when we want to direct debit from a Swiss bank account (debtor address is mandatory when the debtor PSP is located in a non-EEA SEPA country or territory).

So I have extended the BaseTransferInformation class with the attributes townName, postCode, streetName and buildingNumber and their matching getters and setters.

Additionally I have extended the CustomerDirectDebitTransferDomBuilder to add those data to the XML according to the official specification of the ISO 20022 standard.
The specification can be found [here](https://www.iso20022.org/catalogue-messages/iso-20022-messages-archive?search=pain.008.001.02)

Currently this is done for *pain.008.001.02* and the informal *pain.008.003.02*.

A unit test that makes sure that the addition corresponds to the XSD was added, too.